requirements.txt
fastapi[all]
openai
python-dotenv
pydantic==1.*
langchain
faiss-cpu
tiktoken
langchain-community
uvicorn